+++ index.ts
@@ -2,25 +2,13 @@ import {cyan, red, yellow} from 'chalk';
 
 import amqplib = require('amqplib/callback_api');
 import objectAssign = require('object-assign');
-import shortId = require('shortid');
+import {v4} from 'node-uuid';
 import Bluebird = require('bluebird');
 
-const debug = require('debug')('rabbitr');
+import {initWhitelist, shouldSkipSubscribe, log} from './lib/debug';
+import {stringify, stringifyError, parse, parseError} from './lib/serialization';
 
 const DEFAULT_RPC_EXPIRY = 15000; // 15 seconds
-
-// istanbul ignore next
-if (parseFloat(process.version.match(/^v(\d+\.\d+)/)[1]) < 0.4) {
-  // Monkey-patch :(
-  // https://github.com/nodejs/node-v0.x-archive/issues/5110
-  Buffer.prototype.toJSON = function() {
-    return { type: 'Buffer', data: Array.prototype.slice.call(this, 0) };
-  };
-}
-
-function stringify(obj: any): string {
-  return JSON.stringify(obj);
-}
 
 function maybeFromCallback<T>(fn: ((done: Rabbitr.Callback<T>) => void) | (() => PromiseLike<T>)): Bluebird<T> {
   let callback: Rabbitr.Callback<T>;
@@ -31,25 +19,6 @@ function maybeFromCallback<T>(fn: ((done: Rabbitr.Callback<T>) => void) | (() =>
     return Bluebird.resolve(val);
   }
   return promise;
-}
-
-// helper function to properly stringify an error object
-function stringifyError(err, filter?, space?) {
-  var plainObject = {
-    stack: err.stack,
-  };
-  Object.getOwnPropertyNames(err).forEach(function(key) {
-    plainObject[key] = err[key];
-  });
-  return JSON.stringify(plainObject, filter, space);
-}
-
-function parse(json: string): any {
-  return JSON.parse(json, function(key, value) {
-    return value && value.type === 'Buffer'
-      ? new Buffer(value.data)
-      : value;
-  });
 }
 
 class TimeoutError extends Error {
@@ -128,11 +97,21 @@ class Rabbitr {
 
   protected connection: amqplib.Connection;
 
+  /**
+   * An array of channel names used for debug mode. If this value is set, calls to
+   *   #subscribe on channels not in this list will be ignored.
+   *
+   * This option is set from the environment variable `RABBITR_DEBUG`.
+   */
+  private debugChannelsWhitelist: string[] | void;
+
   constructor(opts: Rabbitr.IOptions) {
     if (!HAS_WARNED_ABOUT_V8_BREAKING_CHANGE) {
       console.warn('Rabbitr has a major breaking change in version 8 - rpcListener queues are no longer durable. You will need to remove all rpcListener queues from RabbitMQ during deployment.')
-      HAS_WARNED_ABOUT_V8_BREAKING_CHANGE =  true;
+      HAS_WARNED_ABOUT_V8_BREAKING_CHANGE = true;
     }
+
+    this.debugChannelsWhitelist = initWhitelist();
 
     this.opts = objectAssign(<Rabbitr.IOptions>{
       url: '',
@@ -165,9 +144,9 @@ class Rabbitr {
   private connectionPromise: Bluebird<amqplib.Connection>;
 
   private _connect() {
-    debug('#connect');
+    log('#connect');
 
-    debug(`using connection url ${yellow(this.opts.url)}`);
+    log(`using connection url ${yellow(this.opts.url)}`);
 
     this.connectionPromise = Bluebird.fromCallback<amqplib.Connection>(callback =>
       amqplib.connect(this.opts.url, this.opts.connectionOpts, callback)
@@ -201,7 +180,7 @@ class Rabbitr {
           this.connection = conn;
           this.connected = true;
 
-          debug('ready');
+          log('ready');
           return maybeFromCallback<void>(this.opts.setup || (() => Bluebird.resolve()))
             .then(() => {
               this.ready = true;
@@ -216,11 +195,11 @@ class Rabbitr {
   }
 
   // istanbul ignore next
-  public whenReady(callback: Function) {
-    return this.connectionPromise.then(() => void 0).asCallback(callback);
+  public whenReady(callback?: () => void): Bluebird<void> {
+    return this.connectionPromise.then<void>(() => void 0).asCallback(callback);
   }
 
-  private _formatName(name: string) {
+  private _formatName(name: string): string {
     // istanbul ignore next
     if (this.opts.queuePrefix) {
       name = this.opts.queuePrefix + '.' + name;
@@ -231,14 +210,14 @@ class Rabbitr {
 
   // method to destroy anything for this instance of rabbitr
   destroy(cb?: Rabbitr.ErrorCallback): Bluebird<void> {
-    debug('destroying');
+    log('destroying');
     return Bluebird.each(this._openChannels, channel => {
       return Bluebird.fromCallback(callback => channel.close(callback))
         .then(
-          () => debug('channel closed'),
+          () => log('channel closed'),
           err => {
             // istanbul ignore next
-            debug('Error while closing connection', err);
+            log('Error while closing connection', err);
             throw err;
           }
         );
@@ -247,13 +226,13 @@ class Rabbitr {
         this.connection.close(callback)
       ).then(
         () => {
-          debug('connection closed');
+          log('connection closed');
           this.removeAllListeners();
           this.connectionPromise = null;
         },
         err => {
         // istanbul ignore next
-          debug('Error while closing connection', err);
+          log('Error while closing connection', err);
           throw err;
         }
       );
@@ -269,12 +248,12 @@ class Rabbitr {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
-      return this.whenReady(() =>
+      return this.whenReady().then(() =>
         this.send(topic, data, cb, opts)
       );
     }
 
-    debug(yellow('send'), topic, data, opts);
+    log(yellow('send'), topic, data, opts);
 
     return Bluebird.fromCallback(callback =>
       this._publishChannel.assertExchange(this._formatName(topic), 'topic', {}, callback)
@@ -301,14 +280,21 @@ class Rabbitr {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
-      return this.whenReady(() =>
+      return this.whenReady().then(() =>
         this.subscribe(topic, opts, cb)
       );
     }
 
     const options: Rabbitr.ISubscribeOptions = opts;
 
-    debug(cyan('subscribe'), topic, options);
+    log(cyan('subscribe'), topic, options);
+
+    if (
+      shouldSkipSubscribe(this.debugChannelsWhitelist, topic)
+    ) {
+      log(red('skipped'), cyan('subscribe'), topic);
+      return Bluebird.resolve();
+    }
 
     return Bluebird.fromCallback<amqplib.Channel>(callback =>
       this.connection.createChannel(callback)
@@ -333,7 +319,7 @@ class Rabbitr {
             data = parse(data);
           }
 
-          debug(`got ${cyan(topic)}`, data);
+          log(`got a new message on ${cyan(topic)}`, data);
 
           const messageAcknowledgement = new Bluebird((ack: () => void, reject) => {
             const message: Rabbitr.IMessage<TMessage> = {
@@ -359,12 +345,12 @@ class Rabbitr {
           }).then(
             // acknowledged
             () => {
-              debug(`acknowledging message ${cyan(topic)}`, data);
+              log(`acknowledging message ${cyan(topic)}`, data);
               channel.ack(msg);
             },
             // rejected
             error => {
-              debug(`rejecting message ${cyan(topic)}`, data, error);
+              log(`rejecting message ${cyan(topic)}`, data, error);
               console.error(error && error.stack || error);
               channel.nack(msg);
             }
@@ -402,12 +388,12 @@ class Rabbitr {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
-      return this.whenReady(() =>
+      return this.whenReady().then(() =>
         this.bindExchangeToQueue(exchange, queue, cb)
       );
     }
 
-    debug(cyan('bindExchangeToQueue'), exchange, queue);
+    log(cyan('bindExchangeToQueue'), exchange, queue);
 
     return Bluebird.fromCallback<amqplib.Channel>(callback =>
       this.connection.createChannel(callback)
@@ -438,14 +424,14 @@ class Rabbitr {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
-      return this.whenReady(() =>
+      return this.whenReady().then(() =>
         this.setTimer(topic, uniqueID, data, ttl, cb)
       );
     }
 
     var timerQueue = this._timerQueueName(topic, uniqueID);
 
-    debug(yellow('setTimer'), topic, uniqueID, data);
+    log(yellow('setTimer'), topic, uniqueID, data);
 
     return Bluebird.fromCallback(callback =>
       this._timerChannel.assertQueue(this._formatName(timerQueue), {
@@ -472,14 +458,14 @@ class Rabbitr {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
-      return this.whenReady(() =>
+      return this.whenReady().then(() =>
         this.clearTimer(topic, uniqueID, cb)
       );
     }
 
     var timerQueue = this._timerQueueName(topic, uniqueID);
 
-    debug(yellow('clearTimer'), timerQueue);
+    log(yellow('clearTimer'), timerQueue);
 
     return Bluebird.fromCallback(callback =>
       this._timerChannel.deleteQueue(timerQueue, {}, callback)
@@ -492,7 +478,7 @@ class Rabbitr {
   }
 
   private _getTempQueue(queueName: string, channel: amqplib.Channel) {
-    debug(`creating temp queue ${cyan(queueName)}`)
+    log(`creating temp queue ${cyan(queueName)}`)
     return Bluebird.fromCallback<amqplib.Replies.AssertQueue>(callback =>
       channel.assertQueue(queueName, {
         exclusive: true,
@@ -500,17 +486,16 @@ class Rabbitr {
         durable: false,
       }, callback)
     ).disposer(() => {
-      debug(`deleting temp queue ${cyan(queueName)}`);
-      return Bluebird.fromCallback<void>(callback =>
+      log(`deleting temp queue ${cyan(queueName)}`);
+      return Bluebird.fromCallback<any>(callback =>
         // delete the return queue and close exc channel
-        channel.deleteQueue(queueName) && 0 || callback(null)
-        // FIXME - callback isn't being called here.
+        channel.deleteQueue(queueName, {}, callback)
       ).catch(error => {
         // istanbul ignore next
         console.log(`rabbitr temp queue '${cyan(queueName)}' cleanup exception`, error && error.stack || error);
         throw error;
       }).then(() => {
-        debug(`deleted temp queue ${cyan(queueName)}`);
+        log(`deleted temp queue ${cyan(queueName)}`);
       });
     });
   }
@@ -520,11 +505,11 @@ class Rabbitr {
   rpcExec<TInput, TOutput>(topic: string, data: TInput, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
   rpcExec<TInput, TOutput>(topic: string, data: TInput, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput>;
 
-  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput> {
+  rpcExec<TInput, TOutput>(topic: string, data: TInput, opts?: Rabbitr.IRpcExecOptions, cb?: Rabbitr.Callback<TOutput>): Bluebird<TOutput> {
     // istanbul ignore next
     if (!this.connectionPromise.isFulfilled()) {
       // delay until ready
-      return this.whenReady(() =>
+      return this.whenReady().then(() =>
         this.rpcExec<TInput, TOutput>(topic, data, opts, cb)
       );
     }
@@ -533,13 +518,13 @@ class Rabbitr {
     if ('function' === typeof opts) {
       // shift arguments
       cb = <Rabbitr.Callback<TOutput>>opts;
-      opts = <Rabbitr.IRpcExecOptions>{};
+      opts = null;
     }
 
     // this will send the data down the topic and then open up a unique return queue
     const rpcQueue = this._rpcQueueName(topic);
 
-    const unique = shortId.generate() + '_' + ((Math.round(new Date().getTime() / 1000) + '').substr(5));
+    const unique = v4() + '_' + ((Math.round(new Date().getTime() / 1000) + '').substr(5));
     const returnQueueName = `${rpcQueue}.return.${unique}`;
 
     const now = new Date().getTime();
@@ -550,9 +535,9 @@ class Rabbitr {
     const queueDisposer = this._getTempQueue(this._formatName(returnQueueName), channel);
 
     return Bluebird.using(queueDisposer, () => {
-      debug(`using rpc return queue ${cyan(returnQueueName)}`);
+      log(`using rpc return queue ${cyan(returnQueueName)}`);
 
-      const timeoutMS = (opts.timeout || this.opts.defaultRPCExpiry || DEFAULT_RPC_EXPIRY) * 1;
+      const timeoutMS = (opts && opts.timeout || this.opts.defaultRPCExpiry || DEFAULT_RPC_EXPIRY) * 1;
 
       const replyQueue = this._formatName(returnQueueName);
 
@@ -561,7 +546,7 @@ class Rabbitr {
         .fromCallback<amqplib.Message>(callback => { replyCallback = callback; });
       let gotReply = function(msg) {
         if (!msg) return;
-        debug(`got rpc reply on ${cyan(replyQueue)}`);
+        log(`got rpc reply on ${cyan(replyQueue)}`);
 
         replyCallback(null, msg);
       }
@@ -579,7 +564,7 @@ class Rabbitr {
           expiration: now + timeoutMS,
         };
 
-        debug('sending rpc request');
+        log('sending rpc request');
         this._publishChannel.sendToQueue(this._formatName(rpcQueue), new Buffer(stringify(request)), {
           contentType: 'application/json',
           expiration: `${timeoutMS}`,
@@ -602,18 +587,11 @@ class Rabbitr {
             const response: TOutput = data.response;
 
             if (error) {
-              error = JSON.parse(error);
               if (data.isError) {
-                var err: any = new Error(error.message);
-                Object.keys(error).forEach(function(key) {
-                  if (err[key] !== error[key]) {
-                    err[key] = error[key];
-                  }
-                });
-                error = err;
+                throw parseError(error);
+              } else {
+                throw parse(error);
               }
-
-              throw error;
             }
 
             return response;
@@ -625,8 +603,7 @@ class Rabbitr {
             throw error;
           }
         ).catch(TimeoutError, error => {
-          debug(`request timeout firing for ${rpcQueue} to ${returnQueueName}`);
-          gotReply({}); // clean up so we don't have any unresolved promises left
+          log(`request timeout firing for ${rpcQueue} to ${returnQueueName}`);
           throw error;
         });
       });
@@ -658,10 +635,7 @@ class Rabbitr {
 
     var rpcQueue = this._rpcQueueName(topic);
 
-    (<any>opts).skipMiddleware = true;
-    (<any>opts).durable = false;
-
-    debug(`has rpcListener for ${topic}`);
+    log(`has rpcListener for ${topic}`);
 
     this.on(rpcQueue, (envelope: Rabbitr.IEnvelopedMessage<TInput>) => {
       const dataEnvelope = envelope.data;
@@ -705,15 +679,15 @@ class Rabbitr {
           })
           .then<any>( // sanitize errors
             response => {
-              debug(`${yellow('rpcListener')} responding to topic ${cyan(topic)} with`, response);
+              log(`${yellow('rpcListener')} responding to topic ${cyan(topic)} with`, response);
               return {response};
             }, error => {
-              debug(`${yellow('rpcListener')} on topic ${cyan(topic)} ${red('hit error')}`, error);
+              log(`${yellow('rpcListener')} on topic ${cyan(topic)} ${red('hit error')}`, error);
 
               var isError = error instanceof Error;
               var errJSON = isError ?
                 stringifyError(error) :
-                JSON.stringify(error);
+                stringify(error);
 
               return {
                 error: errJSON,
@@ -733,7 +707,12 @@ class Rabbitr {
       });
     });
 
-    return this.subscribe(rpcQueue, opts).asCallback(callback);
+    return this.subscribe(rpcQueue,
+      objectAssign({}, opts, {
+        skipMiddleware: true,
+        durable: false,
+      })
+    ).asCallback(callback);
   }
 
   // message middleware support