+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## next
+- fixed a bug that caused the return value of `rpcExec` and other methods to always be `undefined` if called before the connection was established.
+- introduced debug mode (via `RABBITR_DEBUG` env variable. check README.md)
+
+## 8.3.0
+- use node-uuid instead of shortid
+
 ## 8.2.3
 - critical bugfix in `subscribe`/`rpcListener` and durable queues